Basic Bot Functionality and Beginning UI
NOTE: To run this code the LWJGL3 library needs to be downloaded and
added to the Java build path, you can download the library here:
https://www.lwjgl.org/download

+ Bots can now play the game (although its a little buggy)
+ Very barebones interface that only shows the color of the card
currently at the top of the discard pile and changes when the bots play
a new card. Will be incorporating the new textures this weekend.
+ Some great new textures to use for the game.
+ Game states, so the game can react differently depending on where it
is (main menu, in-game, etc.)
+ Game is now capped at 60 fps and 60 updates per second thanks to some
great open source code.